requirements.txt
fastapi
uvicorn
pinecone
openai
langchain
langgraph
llama-index
cohere
pymupdf
langchain-text-splitters
peft
transformers
datasets
accelerate
mlflow
ragas
streamlit
python-multipart
python-dotenv
tqdm